计算机程序
增加了小数点

+++ CACULATER927/ViewController.swift
@@ -10,7 +10,7 @@ import UIKit
 
 class ViewController:
 UIViewController {
-    var temp = 0
+    var temp = 0.00
     var flag = 0
     @IBOutlet weak var CACULATERDisplay: UITextField!
     
@@ -51,52 +51,48 @@ UIViewController {
     }
     
     @IBAction func buttonCaculater(_ sender: Any) {
-        if flag==1
-        {
-            var sum = 0
-        sum = temp + Int(CACULATERDisplay.text!)!
+        if flag==1{
+            var sum = 0.00
+        sum = temp + Double(CACULATERDisplay.text!)!
         CACULATERDisplay.text = "\(sum)"
         }
-        if flag==2
-        {
-            var subtraction = 0
-            subtraction = temp - Int(CACULATERDisplay.text!)!
+        if flag==2{
+            var subtraction = 0.00
+            subtraction = temp - Double(CACULATERDisplay.text!)!
             CACULATERDisplay.text = "\(subtraction)"
             
         }
-        if flag==3
-        {
-            var multiplication = 0
-            multiplication = temp * Int(CACULATERDisplay.text!)!
+        if flag==3{
+            var multiplication = 0.00
+            multiplication = temp * Double(CACULATERDisplay.text!)!
             CACULATERDisplay.text = "\(multiplication)"
         }
-        if flag==4
-        {
-            var division = 0
-            division = temp / Int(CACULATERDisplay.text!)!
+        if flag==4{
+            var division = 0.00
+            division = temp / Double(CACULATERDisplay.text!)!
             CACULATERDisplay.text = "\(division)"        }
     }
     
     @IBAction func buttonAdd(_ sender: Any) {
-        temp = Int(CACULATERDisplay.text!)!
+        temp = Double(CACULATERDisplay.text!)!
         CACULATERDisplay.text = ""
         flag=1
     }
     
     @IBAction func buttonsub(_ sender: Any) {
-        temp = Int(CACULATERDisplay.text!)!
+        temp = Double(CACULATERDisplay.text!)!
         CACULATERDisplay.text = ""
         flag=2
     }
     
     @IBAction func buttonmul(_ sender: Any) {
-        temp = Int(CACULATERDisplay.text!)!
+        temp = Double(CACULATERDisplay.text!)!
         CACULATERDisplay.text = ""
         flag=3
     }
     
     @IBAction func buttondis(_ sender: Any) {
-        temp = Int(CACULATERDisplay.text!)!
+        temp = Double(CACULATERDisplay.text!)!
         CACULATERDisplay.text = ""
         flag=4
     }
